requirements.txt
# 深度学习框架
pytorch>=1.13.0
torchvision>=0.14.0
torchaudio>=0.13.0

# 数据处理
numpy>=1.22.0
scipy>=1.8.0
scikit-image>=0.19.0

# 图像处理
Pillow>=9.0.0
opencv-python>=4.5.0

# HuggingFace生态
transformers>=4.24.0
diffusers>=0.12.0
accelerate>=0.14.0

# 可视化和监控
matplotlib>=3.5.0
wandb>=0.13.0

# 硬件监控
pynvml>=11.4.0
psutil>=5.9.0

# 其他工具
zlib
pickle
math
random
os
sys
socket
struct
copy
time
shutil
pathlib
itertools
chain
collections
